lynx-ai/pyproject.toml
[project]
name = "lynx-ai"
version = "0.1.0"

# Protocol Versions
LYNX_PROTOCOL_VERSION = "0.1.0"
MCP_TOOLSET_VERSION = "0.1.0"
description = "Kernel-Governed Artificial Intelligence for AI-BOS NexusCanon"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp-agent[openai]>=0.2.5",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "supabase>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "respx>=0.20.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --tb=short"

[project.scripts]
lynx = "lynx.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

